#1 [SETTING] : remove base activity
