+++ feature/main/src/main/java/com/teamwable/main/MainActivity.kt
@@ -1,19 +1,33 @@
 package com.teamwable.main
 
+import android.annotation.SuppressLint
+import android.content.pm.ActivityInfo
+import android.os.Bundle
 import androidx.activity.viewModels
+import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import com.teamwable.main.databinding.ActivityMainBinding
-import com.teamwable.ui.base.BindingActivity
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
 import timber.log.Timber
 
 @AndroidEntryPoint
-class MainActivity : BindingActivity<ActivityMainBinding>(ActivityMainBinding::inflate) {
+class MainActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityMainBinding
     private val viewModel: DummyViewModel by viewModels()
 
-    override fun initView() {
+    @SuppressLint("SourceLockedOrientationActivity")
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        initView()
+        requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT
+    }
+
+    private fun initView() {
         viewModel.loadDummy()
         lifecycleScope.launch {
             viewModel.uiState
